+++ yandex/base_alb_test.go
@@ -2,10 +2,245 @@ package yandex
 
 import (
 	"fmt"
+	"github.com/fatih/structs"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 	"github.com/yandex-cloud/go-genproto/yandex/cloud/apploadbalancer/v1"
 )
+
+const albDefaultSni = "tf-test-tls"
+const albDefaultValidationContext = "tf-test-validation-context"
+const albDefaultBackendWeight = "1"
+const albDefaultPanicThreshold = "50"
+const albDefaultLocalityPercent = "35"
+const albDefaultTimeout = "300s"
+const albDefaultInterval = "1m500s"
+const albDefaultStrictLocality = "true"
+const albDefaultServiceName = "true"
+const albDefaultHttp2 = "true"
+const albDefaultHost = "tf-test-host"
+const albDefaultPath = "tf-test-path"
+const albDefaultPort = "3"
+const albDefaultSend = "tf-test-send"
+const albDefaultReceive = "tf-test-receive"
+const albDefaultDescription = "alb-bg-description"
+
+type resourceALBBackendGroupInfo struct {
+	IsHttpBackend bool
+	IsGrpcBackend bool
+	IsHttpCheck   bool
+	IsGrpcCheck   bool
+	IsStreamCheck bool
+	IsDataSource  bool
+
+	BaseTemplate string
+
+	TGName string
+	BGName string
+
+	BGDescription        string
+	TlsSni               string
+	TlsValidationContext string
+	BackendWeight        string
+	PanicThreshold       string
+	LocalityPercent      string
+	StrictLocality       string
+	Timeout              string
+	Interval             string
+	ServiceName          string
+	Http2                string
+	Host                 string
+	Path                 string
+	Port                 string
+	Receive              string
+	Send                 string
+}
+
+func albBackendGroupInfo() resourceALBBackendGroupInfo {
+	res := resourceALBBackendGroupInfo{
+		IsHttpBackend:        false,
+		IsGrpcBackend:        false,
+		IsHttpCheck:          false,
+		IsGrpcCheck:          false,
+		IsStreamCheck:        false,
+		IsDataSource:         false,
+		BaseTemplate:         testAccALBBaseTemplate(acctest.RandomWithPrefix("tf-instance")),
+		TGName:               acctest.RandomWithPrefix("tf-tg"),
+		BGName:               acctest.RandomWithPrefix("tf-bg"),
+		BGDescription:        albDefaultDescription,
+		TlsSni:               albDefaultSni,
+		TlsValidationContext: albDefaultValidationContext,
+		BackendWeight:        albDefaultBackendWeight,
+		PanicThreshold:       albDefaultPanicThreshold,
+		LocalityPercent:      albDefaultLocalityPercent,
+		StrictLocality:       albDefaultStrictLocality,
+		Timeout:              albDefaultTimeout,
+		Interval:             albDefaultInterval,
+		ServiceName:          albDefaultServiceName,
+		Http2:                albDefaultHttp2,
+		Host:                 albDefaultHost,
+		Path:                 albDefaultPath,
+		Port:                 albDefaultPort,
+		Receive:              albDefaultReceive,
+		Send:                 albDefaultSend,
+	}
+
+	return res
+}
+
+const albBackendGroupConfigTemplate = `
+{{ if .IsDataSource }}
+data "yandex_alb_backend_group" "test-bg-ds" {
+  name = yandex_alb_backend_group.test-bg.name
+}		
+{{ end }}
+resource "yandex_alb_backend_group" "test-bg" {
+  name        = "{{.BGName}}"
+  description = "{{.BGDescription}}"
+
+  labels = {
+    tf-label    = "tf-label-value"
+    empty-label = ""
+  }
+  {{ if .IsHttpBackend }}
+  http_backend {
+    name             = "test-http-backend"
+    weight           = {{.BackendWeight}}
+    port             = {{.Port}}
+    target_group_ids = ["${yandex_alb_target_group.test-target-group.id}"]
+    tls {
+      sni = "{{.TlsSni}}"
+      validation_context {
+        trusted_ca_bytes = "{{.TlsValidationContext}}"
+      }
+    }
+    load_balancing_config {
+      panic_threshold                = {{.PanicThreshold}}
+      locality_aware_routing_percent = {{.LocalityPercent}}
+      strict_locality                = {{.StrictLocality}}
+    }
+    {{ if .IsGrpcCheck }}
+    healthcheck {
+      timeout  = "{{.Timeout}}"
+      interval = "{{.Interval}}"
+      grpc_healthcheck {
+        service_name = "{{.ServiceName}}"
+      }
+    }
+    {{end}}
+    {{ if .IsStreamCheck }}
+    healthcheck {
+      timeout  = "{{.Timeout}}"
+      interval = "{{.Interval}}"
+      stream_healthcheck {
+        receive = "{{.Receive}}"
+        send    = "{{.Send}}"
+      }
+    }
+    {{end}}
+    {{ if .IsHttpCheck }}
+    healthcheck {
+      timeout = "{{.Timeout}}"
+      interval = "{{.Interval}}"
+      http_healthcheck {
+        host  = "{{.Host}}"
+        path  = "{{.Path}}"
+        http2 = "{{.Http2}}"
+      }
+    }
+    {{end}}
+    http2 = "{{.Http2}}"
+  }
+  {{end}}
+  {{ if .IsGrpcBackend }}
+  grpc_backend {
+    name             = "test-grpc-backend"
+    weight           = {{.BackendWeight}}
+    port             = {{.Port}}
+    target_group_ids = ["${yandex_alb_target_group.test-target-group.id}"]
+    tls {
+      sni = "{{.TlsSni}}"
+      validation_context {
+        trusted_ca_bytes = "{{.TlsValidationContext}}"
+      }
+    }
+    load_balancing_config {
+      panic_threshold                = {{.PanicThreshold}}
+      locality_aware_routing_percent = {{.LocalityPercent}}
+      strict_locality                = {{.StrictLocality}}
+    }
+    {{ if .IsGrpcCheck }}
+    healthcheck {
+      timeout  = "{{.Timeout}}"
+      interval = "{{.Interval}}"
+      grpc_healthcheck {
+        service_name = "{{.ServiceName}}"
+      }
+    }
+    {{end}}
+    {{ if .IsStreamCheck }}
+    healthcheck {
+      timeout  = "{{.Timeout}}"
+      interval = "{{.Interval}}"
+      stream_healthcheck {
+        receive = "{{.Receive}}"
+        send    = "{{.Send}}"
+      }
+    }
+    {{end}}
+    {{ if .IsHttpCheck }}
+    healthcheck {
+      timeout  = "{{.Timeout}}"
+      interval = "{{.Interval}}"
+      http_healthcheck {
+        host  = "{{.Host}}"
+        path  = "{{.Path}}"
+        http2 = "{{.Http2}}"
+      }
+    }
+    {{end}}
+  }
+  {{end}}
+}
+{{ if or .IsHttpBackend .IsGrpcBackend }}
+resource "yandex_alb_target_group" "test-target-group" {
+  name		= "{{.TGName}}"
+
+  target {
+	subnet_id	= "${yandex_vpc_subnet.test-subnet.id}"
+	ip_address	= "${yandex_compute_instance.test-instance-1.network_interface.0.ip_address}"
+  }
+
+  target {
+	subnet_id	= "${yandex_vpc_subnet.test-subnet.id}"
+	ip_address	= "${yandex_compute_instance.test-instance-2.network_interface.0.ip_address}"
+  }
+}
+{{ end }}
+
+{{.BaseTemplate}}
+`
+
+func testALBBackendGroupConfig_basic(in resourceALBBackendGroupInfo) string {
+	m := structs.Map(in)
+	config := templateConfig(albBackendGroupConfigTemplate, m)
+	return config
+}
+
+func testAccCheckALBBackendGroupValues(bg *apploadbalancer.BackendGroup, expectedHttpBackends, expectedGrpcBackends bool) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		if (bg.GetHttp() != nil) != expectedHttpBackends {
+			return fmt.Errorf("invalid presence or absence of http backend Application Backend Group %s", bg.Name)
+		}
+
+		if (bg.GetGrpc() != nil) != expectedGrpcBackends {
+			return fmt.Errorf("invalid presence or absence of grpc backend Application Backend Group %s", bg.Name)
+		}
+
+		return nil
+	}
+}
 
 func testAccALBGeneralTGTemplate(tgName, tgDesc, baseTemplate string, targetsCount int, isDataSource bool) string {
 	targets := make([]string, targetsCount)